+++ Snippets/Snippets_4/Errors/ConfigErrorQueue.cs
@@ -1,0 +1,15 @@
+﻿using NServiceBus.Config;
+using NServiceBus.Config.ConfigurationSource;
+
+#region ErrorQueueConfiguration
+class ConfigErrorQueue : IProvideConfiguration<MessageForwardingInCaseOfFaultConfig>
+{
+    public MessageForwardingInCaseOfFaultConfig GetConfiguration()
+    {
+        return new MessageForwardingInCaseOfFaultConfig
+        {
+            ErrorQueue = "error"
+        };
+    }
+}
+#endregion
