+++ Snippets/Snippets_5/Errors/ConfigureFirstLevelRetries.cs
@@ -1,0 +1,15 @@
+﻿using NServiceBus.Config;
+using NServiceBus.Config.ConfigurationSource;
+
+#region FlrConfiguration
+class ConfigureFirstLevelRetries : IProvideConfiguration<TransportConfig>
+{
+    public TransportConfig GetConfiguration()
+    {
+        return new TransportConfig
+        {
+            MaxRetries = 2
+        };
+    }
+}
+#endregion
